feat(docs): replace script pipeline with consolidated sdd-content Docusaurus plugin
Replaces the 9-file scripts/ approach in the scaffold template with a single
consolidated Docusaurus plugin (plugins/sdd-content/index.js) that:
- Handles all transforms (ADRs, OpenSpecs, index pages, graph page) natively
- Implements getPathsToWatch() for hot reload without chokidar-cli + concurrently
- Uses lib-artifact-transforms for YAML frontmatter parsing
- Generates command tiles page from skills/_index.json + SKILL.md frontmatter
  (opt-in via skillsDir option; silently skipped when no manifest is present)

Adds CommandTile component (template + MDXComponents registration) and command-tile
CSS. Closes superseded PR #165 — command tile generation is now part of the plugin.

Governing: ADR-0004 (Docusaurus docs generation), ADR-0006 (upgrade lifecycle),
           ADR-0029 (skill autogeneration), SPEC-0021 (commands migration)
