--- templates/docusaurus/scripts/transform-adrs.js
@@ -1,220 +1,0 @@
-#!/usr/bin/env node
-/**
- * Transform ADRs for Docusaurus
- *
- * Reads ADR files from docs/adrs/ and writes them to docs-generated/decisions/
- * with Docusaurus frontmatter, RFC 2119 keyword highlighting, and cross-references.
- */
-
-const fs = require('fs');
-const path = require('path');
-const { escapeMdxUnsafe } = require('./mdx-escape');
-const {
-  buildAdrMapping,
-  transformRfc2119Keywords,
-  transformSpecReferences,
-  transformAdrReferences,
-  fixMarkdownLinks,
-} = require('./transform-utils');
-const { getGraph, buildMiniDagSection } = require('./graph-data');
-
-const ADRS_SOURCE = path.join(__dirname, '../../docs/adrs');
-const ADRS_DEST = path.join(__dirname, '../../docs-generated/decisions');
-
-// Lazily-cached artifact graph (per ADR-0023 / SPEC-0018). The first
-// caller in the docs build (whichever transform `build-docs.js` requires
-// first) pays the parse cost; subsequent transforms get the same graph
-// for free.
-const ARTIFACT_GRAPH = getGraph();
-
-// Read baseUrl from docusaurus.config.ts
-const configPath = path.join(__dirname, '../docusaurus.config.ts');
-let BASE_URL = '';
-if (fs.existsSync(configPath)) {
-  const configContent = fs.readFileSync(configPath, 'utf-8');
-  const baseUrlMatch = configContent.match(/baseUrl:\s*['"]([^'"]+)['"]/);
-  BASE_URL = baseUrlMatch ? baseUrlMatch[1].replace(/\/$/, '') : '';
-}
-
-// Spec prefix to emoji and spec path mapping (generated by build-spec-mapping.js)
-let SPEC_MAPPING = {};
-let SPEC_EMOJIS = {};
-try {
-  SPEC_MAPPING = require('../src/data/spec-mapping.json');
-  SPEC_EMOJIS = require('../src/data/spec-emojis.json');
-} catch (e) {
-  // Mapping files not yet generated
-}
-
-const ADR_EMOJI = '\ud83d\udcdd';
-
-const ADR_MAPPING = buildAdrMapping(ADRS_SOURCE);
-
-/** Escape double quotes for YAML frontmatter values */
-function escapeYaml(str) {
-  return str.replace(/\\/g, '\\\\').replace(/"/g, '\\"');
-}
-
-/** Escape special chars for JSX attribute values (double-quoted) */
-function escapeJsxAttr(str) {
-  return str.replace(/&/g, '&amp;').replace(/"/g, '&quot;').replace(/</g, '&lt;').replace(/>/g, '&gt;');
-}
-
-function extractTitle(content) {
-  const match = content.match(/^#\s+(.+)$/m);
-  return match ? match[1].trim() : 'Untitled ADR';
-}
-
-function extractMetadata(content) {
-  let status = null;
-  let date = null;
-  let dm = null;
-
-  // Only parse within YAML frontmatter (between --- delimiters)
-  const fmMatch = content.match(/^---\n([\s\S]*?)\n---/);
-  if (fmMatch) {
-    const fm = fmMatch[1];
-    const statusMatch = fm.match(/^status:\s*"?([^"\n]+)"?/m);
-    const dateMatch = fm.match(/^date:\s*"?([^"\n]+)"?/m);
-    const dmMatch = fm.match(/^decision-makers:\s*"?([^"\n]+)"?/m);
-    if (statusMatch) status = statusMatch[1].trim();
-    if (dateMatch) date = dateMatch[1].trim();
-    if (dmMatch) dm = dmMatch[1].trim();
-  }
-
-  if (!status) status = 'unknown';
-  if (!date) date = 'unknown';
-  if (!dm) dm = 'unknown';
-
-  return { status, date, dm };
-}
-
-function escapeBidirectionalArrows(content) {
-  return content.replace(/<-+>/g, (match) => {
-    return match.replace(/</g, '&lt;').replace(/>/g, '&gt;');
-  });
-}
-
-function transformConsequenceKeywords(content) {
-  const lines = content.split('\n');
-  let inCodeBlock = false;
-
-  return lines.map(line => {
-    const trimmed = line.trimStart();
-    if (/^(`{3,}|~{3,})/.test(trimmed)) { inCodeBlock = !inCodeBlock; return line; }
-    if (inCodeBlock) return line;
-    return line.replace(/^(\s*[\*\-]\s+)(Good|Bad|Neutral|Meh|Okay)(,)/i, (match, prefix, keyword, comma) => {
-      const normalizedKeyword = keyword.charAt(0).toUpperCase() + keyword.slice(1).toLowerCase();
-      const cssClass = normalizedKeyword.toLowerCase();
-      return `${prefix}<span className="consequence-keyword ${cssClass}">${normalizedKeyword}</span>${comma}`;
-    });
-  }).join('\n');
-}
-
-function fixCrossSectionPaths(content) {
-  return content.replace(/\]\(\.\.\/openspec\/specs\//g, '](../specs/');
-}
-
-function transformAdr(srcPath, destPath, fileName) {
-  let content = fs.readFileSync(srcPath, 'utf-8');
-
-  if (fileName === '0000-template.md' || fileName === 'README.md') return;
-
-  const isNumberedAdr = /^(?:ADR-)?\d{4}-/.test(fileName);
-  const title = extractTitle(content);
-  const { status, date, dm } = extractMetadata(content);
-
-  const contentWithoutFrontmatter = content.replace(/^---[\s\S]*?---/, '').trim();
-
-  let escapedContent = fixMarkdownLinks(contentWithoutFrontmatter);
-  escapedContent = fixCrossSectionPaths(escapedContent);
-  escapedContent = escapeBidirectionalArrows(escapedContent);
-  escapedContent = transformRfc2119Keywords(escapedContent);
-  escapedContent = transformSpecReferences(escapedContent, { specMapping: SPEC_MAPPING, specEmojis: SPEC_EMOJIS, baseUrl: BASE_URL });
-  escapedContent = transformAdrReferences(escapedContent, { adrMapping: ADR_MAPPING, adrEmoji: ADR_EMOJI, baseUrl: BASE_URL });
-  escapedContent = transformConsequenceKeywords(escapedContent);
-
-  const slug = fileName.replace(/\.md$/, '');
-
-  let sidebarLabel;
-  if (isNumberedAdr) {
-    const adrNum = fileName.match(/^(?:ADR-)?(\d{4})-/)[1];
-    const titleWithoutAdr = title.replace(/^ADR-\d+:\s*/, '');
-    sidebarLabel = `ADR-${adrNum}: ${titleWithoutAdr}`;
-  } else {
-    sidebarLabel = title;
-  }
-
-  const badgeHeader = isNumberedAdr ? `
-<FieldGroup>
-  <Field label="Status">
-    <StatusBadge status="${escapeJsxAttr(status.toUpperCase())}" />
-  </Field>
-  <Field label="Date">
-    <DateBadge date="${escapeJsxAttr(date)}" />
-  </Field>
-  <Field label="Decision Makers">${escapeJsxAttr(dm)}</Field>
-</FieldGroup>
-` : '';
-
-  const isStricken = ['deprecated', 'superseded'].includes(status.toLowerCase());
-  const sidebarClassName = isStricken ? '\nsidebar_class_name: adr-struck' : '';
-
-  const frontmatter = `---\ntitle: "${escapeYaml(title)}"
-sidebar_label: "${escapeYaml(sidebarLabel)}"
-slug: /decisions/${slug}${sidebarClassName}
----
-${badgeHeader}
-`;
-
-  // Extract canonical artifact ID for the per-page mini-DAG. The
-  // `ADR-` prefix is required (matches `graph-data.js`'s
-  // `adrFileRe`); an unprefixed `0023-foo.md` won't be in the graph
-  // anyway, so there's nothing to render a mini-DAG against.
-  const adrIdMatch = fileName.match(/^(ADR-\d{4})/);
-  const artifactId = adrIdMatch ? adrIdMatch[1] : null;
-  const miniDag = buildMiniDagSection(artifactId, ARTIFACT_GRAPH);
-
-  fs.mkdirSync(path.dirname(destPath), { recursive: true });
-  fs.writeFileSync(destPath, frontmatter + escapeMdxUnsafe(escapedContent) + miniDag);
-}
-
-function main() {
-  console.log('Transforming ADRs...');
-
-  if (!fs.existsSync(ADRS_SOURCE)) {
-    console.log('  No decisions directory found, skipping ADR transform');
-    return;
-  }
-
-  if (fs.existsSync(ADRS_DEST)) {
-    fs.rmSync(ADRS_DEST, { recursive: true });
-  }
-  fs.mkdirSync(ADRS_DEST, { recursive: true });
-
-  // No `link` field here -- the explicit index.mdx written by
-  // generate-index.js (with table + Mermaid hierarchy) owns the
-  // /decisions route. Matches how transform-openspecs.js handles the
-  // /specs category.
-  fs.writeFileSync(path.join(ADRS_DEST, '_category_.json'), JSON.stringify({
-    label: 'Architecture Decisions',
-    position: 2,
-  }, null, 2));
-
-  const files = fs.readdirSync(ADRS_SOURCE);
-  let fileCount = 0;
-
-  for (const file of files) {
-    if (!file.endsWith('.md')) continue;
-    if (file === '0000-template.md' || file === 'README.md') continue;
-
-    const srcPath = path.join(ADRS_SOURCE, file);
-    const destPath = path.join(ADRS_DEST, file.replace(/\.md$/, '.mdx'));
-    transformAdr(srcPath, destPath, file);
-    fileCount++;
-  }
-
-  console.log(`  Transformed ${fileCount} ADR files`);
-}
-
-main();
