--- templates/docusaurus/scripts/transform-openspecs.js
@@ -1,293 +1,0 @@
-#!/usr/bin/env node
-/**
- * Transform OpenSpecs for Docusaurus
- *
- * Reads markdown files from docs/openspec/specs/ and writes them to docs-generated/specs/
- * with Docusaurus frontmatter, RFC 2119 keyword highlighting, and cross-references.
- *
- * Domain ordering is auto-discovered from directory structure (alphabetical).
- */
-
-const fs = require('fs');
-const path = require('path');
-const { escapeMdxUnsafe } = require('./mdx-escape');
-const {
-  buildAdrMapping,
-  transformRfc2119Keywords,
-  transformSpecReferences,
-  transformAdrReferences,
-  fixMarkdownLinks,
-} = require('./transform-utils');
-const { getGraph, buildMiniDagSection } = require('./graph-data');
-
-const SPECS_SOURCE = path.join(__dirname, '../../docs/openspec/specs');
-const SPECS_DEST = path.join(__dirname, '../../docs-generated/specs');
-
-// Read baseUrl from docusaurus.config.ts
-const configPath = path.join(__dirname, '../docusaurus.config.ts');
-let BASE_URL = '';
-if (fs.existsSync(configPath)) {
-  const configContent = fs.readFileSync(configPath, 'utf-8');
-  const baseUrlMatch = configContent.match(/baseUrl:\s*['"]([^'"]+)['"]/);
-  BASE_URL = baseUrlMatch ? baseUrlMatch[1].replace(/\/$/, '') : '';
-}
-
-// Spec prefix to emoji and spec path mapping (generated by build-spec-mapping.js)
-let SPEC_MAPPING = {};
-let SPEC_EMOJIS = {};
-try {
-  SPEC_MAPPING = require('../src/data/spec-mapping.json');
-  SPEC_EMOJIS = require('../src/data/spec-emojis.json');
-} catch (e) {
-  // Mapping files not yet generated
-}
-
-// ADR emoji and source path
-const ADR_EMOJI = '\ud83d\udcdd';
-const ADRS_SOURCE = path.join(__dirname, '../../docs/adrs');
-
-const ADR_MAPPING = buildAdrMapping(ADRS_SOURCE);
-
-// Lazily-cached artifact graph (per ADR-0023 / SPEC-0018). Shared
-// with transform-adrs.js and generate-graph.js via getGraph()'s
-// in-process cache, so the frontmatter corpus is parsed once per
-// docs build regardless of which transform runs first.
-const ARTIFACT_GRAPH = getGraph();
-
-// Domain directory -> canonical SPEC-XXXX id, derived from the graph's
-// spec nodes (which carry their source `dir` alongside the parsed id).
-// design.md files don't carry the SPEC-XXXX in their own H1, so we look
-// up by the directory both spec.md and design.md share.
-const SPEC_DOMAIN_TO_ID = {};
-for (const node of Object.values(ARTIFACT_GRAPH.nodes)) {
-  if (node.kind === 'spec' && node.dir) {
-    SPEC_DOMAIN_TO_ID[node.dir] = node.id;
-  }
-}
-
-/** Escape double quotes for YAML frontmatter values */
-function escapeYaml(str) {
-  return str.replace(/\\/g, '\\\\').replace(/"/g, '\\"');
-}
-
-/** Escape special chars for JSX attribute values (double-quoted) */
-function escapeJsxAttr(str) {
-  return str.replace(/&/g, '&amp;').replace(/"/g, '&quot;').replace(/</g, '&lt;').replace(/>/g, '&gt;');
-}
-
-/**
- * Auto-discover domains and build config from directory structure.
- * Generates labels from directory names: "web-dashboard" -> "Web Dashboard"
- */
-function buildDomainConfig() {
-  const config = {};
-  if (!fs.existsSync(SPECS_SOURCE)) return config;
-
-  const domains = fs.readdirSync(SPECS_SOURCE)
-    .filter(d => fs.statSync(path.join(SPECS_SOURCE, d)).isDirectory())
-    .sort();
-
-  domains.forEach((domain, index) => {
-    // Extract prefix from spec.md if possible
-    let prefix = domain.toUpperCase().replace(/-/g, '').slice(0, 6);
-    const specPath = path.join(SPECS_SOURCE, domain, 'spec.md');
-    if (fs.existsSync(specPath)) {
-      const content = fs.readFileSync(specPath, 'utf-8');
-      const specMatch = content.match(/^#\s+SPEC-\d+:\s+(.+)$/m);
-      if (specMatch) {
-        // Use the spec title as the label
-        const label = specMatch[1].trim();
-        config[domain] = { order: index + 1, label: label };
-        return;
-      }
-    }
-
-    // Fallback: humanize the directory name
-    const label = domain
-      .split('-')
-      .map(w => w.charAt(0).toUpperCase() + w.slice(1))
-      .join(' ');
-    config[domain] = { order: index + 1, label: label };
-  });
-
-  return config;
-}
-
-function extractMetadata(content, fileType) {
-  let status = null;
-  let date = null;
-
-  const fmMatch = content.match(/^---\n([\s\S]*?)\n---/);
-  if (fmMatch) {
-    const fm = fmMatch[1];
-    const statusMatch = fm.match(/^status:\s*(.+)$/m);
-    const dateMatch = fm.match(/^date:\s*(.+)$/m);
-    if (statusMatch) status = statusMatch[1].trim();
-    if (dateMatch) date = dateMatch[1].trim();
-    content = content.slice(fmMatch[0].length).replace(/^\n+/, '');
-  }
-
-  if (!status) status = fileType === 'spec' ? 'active' : 'draft';
-  if (!date) date = 'unknown';
-
-  return { status, date, content };
-}
-
-function extractTitle(content) {
-  const match = content.match(/^#\s+(.+)$/m);
-  return match ? match[1].trim() : 'Untitled';
-}
-
-function fixRelativePaths(content) {
-  content = content.replace(/\]\(\.\.\/\.\.\/\.\.\/decisions\//g, '](../../decisions/');
-  content = content.replace(/\]\(\.\.\/\.\.\/\.\.\/docs\//g, '](../../');
-  return content;
-}
-
-function transformRequirementTables(content) {
-  const tableRegex = /\| ID \| Requirement \|\n\|[-| ]+\|\n((?:\| [A-Z0-9-]+ \| .* \|\n)+)/g;
-
-  return content.replace(tableRegex, (match, rows) => {
-    const rowRegex = /\| ([A-Z0-9-]+) \| (.*) \|/g;
-    let result = '';
-    let rowMatch;
-
-    while ((rowMatch = rowRegex.exec(rows)) !== null) {
-      const id = rowMatch[1].trim();
-      const text = rowMatch[2].trim();
-      result += `<RequirementBox id="${id}">\n\n${text}\n\n</RequirementBox>\n\n`;
-    }
-
-    return result;
-  });
-}
-
-function transformSpec(srcPath, destPath, domain, fileType, domainConfig, flat) {
-  let content = fs.readFileSync(srcPath, 'utf-8');
-  const config = domainConfig[domain] || { order: 99, label: domain };
-
-  const metadata = extractMetadata(content, fileType);
-  content = metadata.content;
-  const title = extractTitle(content);
-
-  content = fixRelativePaths(content);
-  content = fixMarkdownLinks(content);
-
-  if (fileType === 'spec') {
-    content = transformRequirementTables(content);
-  }
-
-  content = transformRfc2119Keywords(content);
-  content = transformSpecReferences(content, { specMapping: SPEC_MAPPING, specEmojis: SPEC_EMOJIS, baseUrl: BASE_URL });
-  content = transformAdrReferences(content, { adrMapping: ADR_MAPPING, adrEmoji: ADR_EMOJI, baseUrl: BASE_URL });
-
-  const sidebarLabel = flat ? config.label : (fileType === 'spec' ? 'Specification' : 'Design');
-  const sidebarPosition = flat ? config.order : (fileType === 'spec' ? 1 : 2);
-
-  const metadataHeader = `
-<FieldGroup>
-  <Field label="Status">
-    <StatusBadge status="${escapeJsxAttr(metadata.status.toUpperCase())}" />
-  </Field>
-  <Field label="Date">
-    <DateBadge date="${escapeJsxAttr(metadata.date)}" />
-  </Field>
-  <Field label="Domain">
-    <DomainBadge domain="${escapeJsxAttr(config.label)}" />
-  </Field>
-</FieldGroup>
-`;
-
-  const frontmatter = `---
-title: "${escapeYaml(title)}"
-sidebar_label: "${escapeYaml(sidebarLabel)}"
-sidebar_position: ${sidebarPosition}
----
-
-${metadataHeader}
-
-`;
-
-  // Both spec.md and design.md describe the same SPEC-XXXX artifact and
-  // get the same neighbor graph; the mini-DAG renders identically on
-  // each, which matches user expectation that "this artifact's
-  // relationships" is an attribute of the artifact, not the page.
-  const artifactId = SPEC_DOMAIN_TO_ID[domain];
-  const miniDag = buildMiniDagSection(artifactId, ARTIFACT_GRAPH);
-
-  fs.mkdirSync(path.dirname(destPath), { recursive: true });
-  fs.writeFileSync(destPath, frontmatter + escapeMdxUnsafe(content) + miniDag);
-}
-
-function generateCategoryJson(destDir, domain, domainConfig) {
-  const config = domainConfig[domain] || { order: 99, label: domain };
-
-  const categoryData = {
-    label: config.label,
-    position: config.order,
-    link: {
-      type: 'generated-index',
-      description: `${config.label} specifications and design documents.`
-    }
-  };
-
-  const categoryPath = path.join(destDir, '_category_.json');
-  fs.writeFileSync(categoryPath, JSON.stringify(categoryData, null, 2));
-}
-
-function main() {
-  console.log('Transforming OpenSpecs...');
-
-  if (!fs.existsSync(SPECS_SOURCE)) {
-    console.log('  No specs directory found, skipping OpenSpec transform');
-    return;
-  }
-
-  // Auto-discover domain config
-  const domainConfig = buildDomainConfig();
-
-  if (fs.existsSync(SPECS_DEST)) {
-    fs.rmSync(SPECS_DEST, { recursive: true });
-  }
-  fs.mkdirSync(SPECS_DEST, { recursive: true });
-
-  fs.writeFileSync(path.join(SPECS_DEST, '_category_.json'), JSON.stringify({
-    label: 'Specifications',
-    position: 1,
-  }, null, 2));
-
-  const domains = fs.readdirSync(SPECS_SOURCE);
-  let fileCount = 0;
-
-  for (const domain of domains) {
-    const domainPath = path.join(SPECS_SOURCE, domain);
-    if (!fs.statSync(domainPath).isDirectory()) continue;
-
-    const hasSpec = fs.existsSync(path.join(domainPath, 'spec.md'));
-    const hasDesign = fs.existsSync(path.join(domainPath, 'design.md'));
-
-    if (!hasSpec && !hasDesign) continue;
-
-    if (hasSpec && hasDesign) {
-      // Both docs: create subdirectory with _category_.json
-      const destDomainPath = path.join(SPECS_DEST, domain);
-      fs.mkdirSync(destDomainPath, { recursive: true });
-      generateCategoryJson(destDomainPath, domain, domainConfig);
-
-      transformSpec(path.join(domainPath, 'spec.md'), path.join(destDomainPath, 'spec.mdx'), domain, 'spec', domainConfig, false);
-      transformSpec(path.join(domainPath, 'design.md'), path.join(destDomainPath, 'design.mdx'), domain, 'design', domainConfig, false);
-      fileCount += 2;
-    } else {
-      // Single doc: emit as flat file (leaf item in sidebar)
-      const file = hasSpec ? 'spec.md' : 'design.md';
-      const fileType = hasSpec ? 'spec' : 'design';
-      const destPath = path.join(SPECS_DEST, `${domain}.mdx`);
-      transformSpec(path.join(domainPath, file), destPath, domain, fileType, domainConfig, true);
-      fileCount++;
-    }
-  }
-
-  console.log(`  Transformed ${fileCount} spec files across ${Object.keys(domainConfig).length} domains`);
-}
-
-main();
