--- templates/docusaurus/src/data/spec-emojis.json
@@ -1,1 +1,0 @@
-{}
