--- templates/docusaurus/src/data/spec-mapping.json
@@ -1,1 +1,0 @@
-{}
